refactor(ui): split single and range date filter bodies
SearchFilterConditionDate had grown to four state variables, a
branching apply handler, and duplicated DatePicker blocks to serve
two distinct input shapes. Splitting the bodies into
SearchFilterDateSingle and SearchFilterDateRange gives each input
shape its own state, validation, and apply handler, so the parent
only owns the condition selection. This keeps the upcoming
relative-age work (ROX-28073 stretch) from growing the same
component further.

Shared date parsing and formatting helpers move to utils/dateInput
so both bodies use identical strict MM/DD/YYYY handling. Behavior
and aria-labels are unchanged; both Cypress component specs pass
without selector changes.
